Select2 has two operation modes: 1) read data from the options of an existing select element -> select2 must attached to an select element 2) handle data them internally (set them on initialization or load them through ajax) -> select2 must be attached to an hideen input element
If you want an entry to be selected the format is different for the cases mentioned above:
1) either a single string if you have a single selection element or a list of strings if you have a multiple selection element
2) either an object {id:<idvalue>, text:<textvalue>} if you have a single sleection element or a list of such objects if you have a multiple selection element

nb: you can have more props in the object, those mentioned are the minimum for the default renderer

To facilitate these options i added a new parameter `ondemand` which covers the 2nd case mentioned above when it's value is True.
The default value is False though thus implementing option 1.